Patient service

+++ PatientService/src/main/java/ba/unsa/etf/nwt/PatientService/repositories/TestItemRepository.java
@@ -3,5 +3,5 @@ package ba.unsa.etf.nwt.PatientService.repositories;
 import ba.unsa.etf.nwt.PatientService.model.TestItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestItemRepository extends JpaRepository<TestItem, Integer> {
+public interface TestItemRepository extends JpaRepository<TestItem, Long> {
 }
