+++ PatientService/src/main/java/ba/unsa/etf/nwt/PatientService/repositories/TestResultRepository.java
@@ -3,5 +3,5 @@ package ba.unsa.etf.nwt.PatientService.repositories;
 import ba.unsa.etf.nwt.PatientService.model.TestResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestResultRepository extends JpaRepository<TestResult, Integer> {
+public interface TestResultRepository extends JpaRepository<TestResult, Long> {
 }
